% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_arima.R
\name{CoefficientsTable}
\alias{CoefficientsTable}
\title{Function to create a table of the estimated model coefficients from a call to CausalArima}
\usage{
CoefficientsTable(x, ...)
}
\arguments{
\item{x}{Object of class \code{cArima}.}

\item{...}{Optional arguments to pass on \code{stargazer()}.}
}
\description{
Function to create a table of the estimated model coefficients from a call to CausalArima
}
\examples{
## Example 2 (weekly data, with predictors)
# Generating a time series of length 800 and a vector of dates
y <- rnorm(800, sd = 1)
dates <- seq.Date(from = as.Date("2005-01-01"), by = "week", length.out = 800)

# Generating predictors
x1 <- rnorm(800, mean = 2, sd = 0.5)
x2 <- rnorm(800, mean = 3, sd = 0.5)
y <- y -2*x1 + x2

# Adding a fictional intervention
int.date <- as.Date("2019-05-11")
horizon <- c(as.Date("2019-12-07"), as.Date("2020-02-15"), as.Date("2020-04-25"))
y.new <- y ; y.new[dates >= int.date] <- y.new[dates >= int.date]*1.40

# Causal effect estimation
start<-as.numeric(strftime(as.Date(dates[1], "\%Y-\%m-\%d"), "\%u"))
ce <- CausalArima(y = ts(y.new, start = start, frequency = 1), auto = TRUE, ic = "aic", dates = dates, xreg = data.frame(x1,x2), int.date = int.date)

# Table of the estimated temporal average effects
CoefficientsTable(ce)

}
