% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_arima.R
\name{impact}
\alias{impact}
\title{Function to create a table of the estimated model coefficients from a call to CausalArima}
\usage{
impact(x, format = "numeric", nboot = 10, horizon = NULL, ...)
}
\arguments{
\item{x}{Object of class \code{cArima}.}

\item{format}{Required format for the table. Possible values in \code{c("numeric", "html", "latex")}.}

\item{horizon}{Optional vector with elements of class Date. If provided, the function returns
the estimated effects at the given time horizons.}

\item{...}{Optional arguments passed to other methods.}

\item{n}{Number of bootstrap simulations.}
}
\value{
A list with the following components:
\item{impact}{Estimated point, cumulative and temporal average effect assuming Normality
             of the error terms.}
\item{impact_boot}{Estimated point, cumulative and temporal average effect by bootstrap.}
\item{arima}{Arima order, coefficient estimates and accuracy measures for the estimated model
             in the pre-intervention period.}
}
\description{
Function to create a table of the estimated model coefficients from a call to CausalArima
}
\examples{
## Example 2 (weekly data, with predictors)
# Generating a time series of length 800 and a vector of dates
y <- rnorm(800, sd = 1)
dates <- seq.Date(from = as.Date("2005-01-01"), by = "week", length.out = 800)

# Generating predictors
x1 <- rnorm(800, mean = 2, sd = 0.5)
x2 <- rnorm(800, mean = 3, sd = 0.5)
y <- y -2*x1 + x2

# Adding a fictional intervention
int.date <- as.Date("2019-05-11")
horizon <- c(as.Date("2019-12-07"), as.Date("2020-02-15"), as.Date("2020-04-25"))
y.new <- y ; y.new[dates >= int.date] <- y.new[dates >= int.date]*1.40

# Causal effect estimation
start<-as.numeric(strftime(as.Date(dates[1], "\%Y-\%m-\%d"), "\%u"))
ce <- CausalArima(y = ts(y.new, start = start, frequency = 1), auto = TRUE, ic = "aic", dates = dates, xreg = data.frame(x1,x2), int.date = int.date)

# Table of the estimated temporal average effects
impact(ce)

}
