% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_arima.R
\name{plot.cArima}
\alias{plot.cArima}
\title{Plotting method for object of class cArima}
\usage{
\method{plot}{cArima}(x, int.date, type = c("forecast", "impact", "residuals"))
}
\arguments{
\item{x}{Object of class \code{cArima}.}

\item{int.date}{Date of the intervention (required for types "forecast" and "impact").}

\item{type}{Character string indicating the plot to be produced. Possible values
in c("forecast", "impact", "residuals").}
}
\description{
Plotting method for object of class cArima
}
\details{
The option 'forecast' plots the observed time series as compared to the forecasted
series. The option 'impact' plots the causal effect at each point in time after the
intervention date, as resulting from the direct comparison between the observed and
forecasted time series. Finally, the option 'residuals' draws the autocorrelation and
partial autocorrelation functions of the residuals of the model fitted to the data
in the pre-intervention period as well as the Normal Q-Q Plot.
}
\examples{
## Example 1 (daily data, no predictors)
# Generating a time series of length 1000 and a vector of dates
y <- 0.5*seq(0.5, 250, by = 0.5) + rnorm(500, sd = 6)
dates <- seq.Date(from = as.Date("2014-01-05"), by = "days", length.out = 500)

# Adding a fictional intervention
int.date <- as.Date("2015-04-01")
horizon <- c(as.Date("2015-04-10"), as.Date("2015-04-20"))
y.new <- y ; y.new[dates >= int.date] <- y.new[dates >= int.date]*1.25

# Causal effect estimation
start<-as.numeric(strftime(as.Date(dates[1], "\%Y-\%m-\%d"), "\%u"))
ce <- CausalArima(y = ts(y.new, start = start, frequency = 1), auto = TRUE, ic = "aic", dates = dates, int.date = int.date)

# Plot
plot(ce, type = "forecast", int.date = int.date)
plot(ce, type = "impact", int.date = int.date)
plot(ce, type = "residuals")

}
