% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_var.R
\name{plot.cVAR}
\alias{plot.cVAR}
\title{Plotting method for object of class cVAR}
\usage{
\method{plot}{cVAR}(x, int.date, type = c("forecast", "impact", "residuals"), ...)
}
\arguments{
\item{x}{Object of class \code{cVAR}.}

\item{int.date}{Date of the intervention (required for types "forecast" and "impact").}

\item{type}{Character string indicating the plot to be produced. Possible values
in c("forecast", "impact", "residuals").}
}
\description{
Plotting method for object of class cVAR
}
\details{
The option 'forecast' plots the observed time series as compared to the forecasted
series. The option 'impact' plots the causal effect at each point in time after the
intervention date, as resulting from the direct comparison between the observed and
forecasted time series. Finally, the option 'residuals' draws the autocorrelation and
partial autocorrelation functions of the residuals of the model fitted to the data
in the pre-intervention period as well as the Normal Q-Q Plot.
}
\examples{
# Work in progress
}
