% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_arima.R
\name{pretty_summary}
\alias{pretty_summary}
\title{Title}
\usage{
pretty_summary(
  y,
  x = x1,
  n = 10,
  beta = 0.05,
  digits = 2L,
  boot = FALSE,
  printing = FALSE
)
}
\arguments{
\item{boot}{}
}
\value{

}
\description{
Title
}
