% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_arima.R
\name{ResultTable}
\alias{ResultTable}
\title{Function to create ready-to-use tables of the estimated causal effects from a call to CausalArima}
\usage{
ResultTable(
  x,
  type = "norm",
  stat = c("tau", "avg", "sum"),
  direction = "b",
  horizon = NULL,
  digits = 2,
  printing = FALSE
)
}
\arguments{
\item{x}{Object of class \code{cArima}.}

\item{type}{Character, indicating whether to produce a table reporting
the estimated standard errors under the Normality assumption
(\code{type = "norm"}) or bootstrapped errors (\code{type = "boot"}).}

\item{stat}{Character, indicating the test statistic to include in the table.
Possible values in \code{c("tau", "sum", "avg")}.}

\item{direction}{Character, for the two-sided test set \code{direction = "b"} (default),
for one-sided tests set \code{direction = "l"} or \code{direction = "r"}.}

\item{horizon}{Optional vector with elements of class \code{Date}. If provided, the function
outputs a table for the required \code{stat} at every given time horizon.
If \code{NULL}, the table refers to the last date of the analysis period.}

\item{digits}{Number of decimal places, defaults to 2.}
}
\description{
Function to create ready-to-use tables of the estimated causal effects from a call to CausalArima
}
\examples{
## Example 2 (weekly data, with predictors)
# Generating a time series of length 800 and a vector of dates
y <- rnorm(800, sd = 1)
dates <- seq.Date(from = as.Date("2005-01-01"), by = "week", length.out = 800)

# Generating predictors
x1 <- rnorm(800, mean = 2, sd = 0.5)
x2 <- rnorm(800, mean = 3, sd = 0.5)
y <- y -2*x1 + x2

# Adding a fictional intervention
int.date <- as.Date("2019-05-11")
horizon <- c(as.Date("2019-12-07"), as.Date("2020-02-15"), as.Date("2020-04-25"))
y.new <- y ; y.new[dates >= int.date] <- y.new[dates >= int.date]*1.40

# Causal effect estimation
start<-as.numeric(strftime(as.Date(dates[1], "\%Y-\%m-\%d"), "\%u"))
ce <- CausalArima(y = ts(y.new, start = start, frequency = 1), auto = TRUE, ic = "aic", dates = dates, xreg = data.frame(x1,x2), int.date = int.date)

# Table of the estimated temporal average effects
ResultTable(ce, type = "norm", horizon = horizon)
}
