% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_arima.R
\name{print.cArima}
\alias{print.cArima}
\title{Print method for object of class 'cArima'}
\usage{
\method{print}{cArima}(x, type = "norm", horizon = NULL, digits = 3)
}
\arguments{
\item{x}{Object of class \code{cArima}.}

\item{type}{Character string indicating the summary to be produced. Possible values
in \code{c("norm", "boot")}. Defaults to \code{"norm"}.}

\item{horizon}{Optional vector with elements of class \code{Date}. If provided, the function
summarizes the point, cumulative and temporal average effects at the given
time horizon(s).}

\item{digits}{Integer, indicating the number of decimal places to show in the output.}
}
\value{

}
\description{
Format and prints the point, cumulative and temporal average effects (standard errors
and critical values) in a nice and clean output.
}
