% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_arima.R
\name{print.cArima}
\alias{print.cArima}
\title{Print method for object of class 'cArima'}
\usage{
\method{print}{cArima}(x, type = "norm", horizon = NULL)
}
\arguments{
\item{x}{Object of class \code{cArima}.}

\item{type}{Character string indicating the summary to be produced. Possible values
in \code{c("norm", "boot")}. Defaults to \code{"norm"}.}

\item{horizon}{Optional vector with elements of class \code{Date}. If provided, the function
summarizes the point, cumulative and temporal average effects at the given
time horizon(s).}
}
\value{
A data frame with as many rows as the dates provided in horizon (if \code{is.null(horizon)},
        a single row corresponding to the last date of the post-intervention period) with the
        following columns:
        \item{tau}{The estimated causal effect at the given time horizon or at the end of analysis
                   period if \code{is.null(horizon)}.}
        \item{pvalue.tau}{Left-sided (\code{.l}), bidirectional (\code{.b}) and right-sided (\code{.r})
                          p-values for \code{tau}.}
        \item{sum.tau}{The estimated cumulative causal effect up to the given time horizon or at the end of analysis
                       period if \code{is.null(horizon)}.}
        \item{pvalue.sum}{Left-sided (\code{.l}), bidirectional (\code{.b}) and right-sided (\code{.r})
                          p-values for \code{sum.tau}.}
        \item{avg.tau}{The estimated temporal average causal effect up to the given time horizon or at the end of analysis
                       period if \code{is.null(horizon)}.}
        \item{pvalue.avg}{Left-sided (\code{.l}), bidirectional (\code{.b}) and right-sided (\code{.r})
                          p-values for \code{avg.tau}.}
        When \code{type = "norm"} additional columns are provided with the estimated standard deviations
        for the point, cumulative and temporal average effects under the assumption of Normally distributed residuals.
}
\description{
Print method for object of class 'cArima'
}
\examples{
## Example 1
# Generating a time series of length 1000 and a vector of dates
y <- 0.5*seq(0.5, 250, by = 0.5) + rnorm(500, sd = 6)
dates <- seq.Date(from = as.Date("2014-01-05"), by = "days", length.out = 500)

# Adding a fictional intervention
int.date <- as.Date("2015-04-01")
horizon <- c(as.Date("2015-04-10"), as.Date("2015-04-20"))
y.new <- y ; y.new[dates >= int.date] <- y.new[dates >= int.date]*1.25

# Causal effect estimation
start<-as.numeric(strftime(as.Date(dates[1], "\%Y-\%m-\%d"), "\%u"))
ce <- CausalArima(y = ts(y.new, start = start, frequency = 1), auto = TRUE, ic = "aic",
                  dates = dates, int.date = int.date, nboot = 1000)

# Print
print(ce, type = "norm")
print(ce, type = "boot", horizon = horizon)

}
